@@ -1,11 +1,10 @@
 module github.com/fjl/geas
 
-go 1.21
+go 1.23
 
 require (
 	github.com/ethereum/go-ethereum v1.12.2
 	golang.org/x/crypto v0.13.0
-	golang.org/x/exp v0.0.0-20230905200255-921286631fa9
 	gopkg.in/yaml.v3 v3.0.1
 )
 
