@@ -23,12 +23,10 @@ import (
 	"errors"
 	"fmt"
 	"io/fs"
-	"math"
-	"math/big"
 	"path"
 	"strings"
 
-	"github.com/fjl/geas/internal/evm"
+	"github.com/fjl/geas/internal/ast"
 )
 
 // Compiler performs the assembling.
@@ -41,65 +39,16 @@ type Compiler struct {
 
 	globals    *globalScope
 	errors     []error
-	macroStack map[*instructionMacroDef]struct{}
-}
-
-type astInstruction interface {
-	astStatement
-	expand(c *Compiler, doc *document, prog []*instruction) ([]*instruction, error)
-}
-
-// instruction is a step of the bytecode program.
-type instruction struct {
-	ast astInstruction
-	op  string
-	doc *document
-
-	// compiler-assigned fields:
-	pc          int    // pc at this instruction
-	pushSize    int    // computed size of push instruction
-	data        []byte // computed argument value
-	argNoLabels bool   // true if arg expression does not contain @label
-}
-
-func newInstruction(doc *document, ast astInstruction, op string) *instruction {
-	return &instruction{doc: doc, ast: ast, op: op}
-}
-
-func isPush(op string) bool {
-	return strings.HasPrefix(op, "PUSH")
-}
-
-func isJump(op string) bool {
-	return strings.HasPrefix(op, "JUMP")
-}
-
-// explicitPushSize returns the declared PUSH size.
-func (inst *instruction) explicitPushSize() (int, bool) {
-	op, ok := inst.ast.(*opcodeInstruction)
-	if ok {
-		return int(op.pushSize) - 1, op.pushSize > 0
-	}
-	return 0, false
-}
-
-// pushArg returns the instruction argument.
-func (inst *instruction) pushArg() astExpr {
-	if !isPush(inst.op) {
-		return nil
-	}
-	op, ok := inst.ast.(*opcodeInstruction)
-	if ok {
-		return op.arg
-	}
-	return nil
+	macroStack map[*ast.InstructionMacroDef]struct{}
+	includes   map[*ast.IncludeSt]*ast.Document
 }
 
 // NewCompiler creates a compiler. The passed file system is used to resolve import file names.
 func NewCompiler(fsys fs.FS) *Compiler {
 	return &Compiler{
 		fsys:        fsys,
-		macroStack:  make(map[*instructionMacroDef]struct{}),
+		macroStack:  make(map[*ast.InstructionMacroDef]struct{}),
+		includes:    make(map[*ast.IncludeSt]*ast.Document),
 		maxIncDepth: 128,
 		maxErrors:   10,
 		usePush0:    true,
@@ -150,8 +99,8 @@ func (c *Compiler) CompileFile(filename string) []byte {
 }
 
 func (c *Compiler) compileSource(filename string, input []byte) []byte {
-	p := newParser(filename, input, c.lexDebug)
-	doc, errs := p.parse()
+	p := ast.NewParser(filename, input, c.lexDebug)
+	doc, errs := p.Parse()
 	for _, err := range errs {
 		c.errors = append(c.errors, err)
 	}
@@ -167,7 +116,7 @@ func (c *Compiler) Errors() []error {
 }
 
 // addError pushes an error to the compiler error list.
-func (c *Compiler) addError(inst astInstruction, err error) {
+func (c *Compiler) addError(inst ast.Statement, err error) {
 	c.errors = append(c.errors, &astError{inst: inst, err: err})
 	if len(c.errors) > c.maxErrors {
 		panic(errCancelCompilation)
@@ -185,7 +134,7 @@ func (c *Compiler) addErrors(errs []error) {
 }
 
 // compile is the toplevel entry point into the compiler.
-func (c *Compiler) compile(doc *document) (output []byte) {
+func (c *Compiler) compile(doc *ast.Document) (output []byte) {
 	prevGlobals := c.globals
 	c.globals = newGlobalScope()
 	defer func() { c.globals = prevGlobals }()
@@ -201,8 +150,11 @@ func (c *Compiler) compile(doc *document) (output []byte) {
 	c.processIncludes(doc, nil)
 
 	// Next, the AST document tree is expanded into a flat list of instructions.
-	var prog []*instruction
-	prog = c.expand(doc, doc.instructions, prog)
+	prog := newCompilerProg(doc)
+	c.expand(doc, prog)
+	if prog.cur != prog.toplevel {
+		panic("section stack was not unwound by expansion")
+	}
 
 	// Expansion of is now done, and all further steps work on prog.
 	e := newEvaluator(c.globals)
@@ -231,163 +183,18 @@ func (c *Compiler) compile(doc *document) (output []byte) {
 	return c.generateOutput(prog)
 }
 
-// expand appends a list of AST instructions to the program.
-func (c *Compiler) expand(doc *document, input []astInstruction, prog []*instruction) []*instruction {
-	for _, inst := range input {
-		newprog, err := inst.expand(c, doc, prog)
-		if err != nil {
-			c.addError(inst, err)
-			continue
-		}
-		prog = newprog
-	}
-	return prog
-}
-
-// expand creates an instruction for the label. For dotted labels, the instruction is
-// empty (i.e. has size zero). For regular labels, a JUMPDEST is created.
-func (li *labelDefInstruction) expand(c *Compiler, doc *document, prog []*instruction) ([]*instruction, error) {
-	if li.global {
-		if err := c.globals.setLabelDocument(li, doc); err != nil {
-			c.addErrors([]error{err})
-		}
-	}
-
-	inst := newInstruction(doc, li, "")
-	if !li.dotted {
-		inst.op = "JUMPDEST"
-	}
-	prog = append(prog, inst)
-	return prog, nil
-}
-
-// expand appends the instruction to a program. This is also where basic validation is done.
-func (op *opcodeInstruction) expand(c *Compiler, doc *document, prog []*instruction) ([]*instruction, error) {
-	opcode := strings.ToUpper(op.tok.text)
-	inst := newInstruction(doc, op, opcode)
-
-	switch {
-	case isPush(opcode):
-		if opcode == "PUSH0" {
-			if op.arg != nil {
-				return nil, ecPushzeroWithArgument
-			}
-			break
-		}
-		if op.arg == nil {
-			return prog, ecPushWithoutArgument
-		}
-
-	case isJump(opcode):
-		if err := c.validateJumpArg(doc, op.arg); err != nil {
-			return nil, err
-		}
-		// 'JUMP @label' instructions turn into 'PUSH @label' + 'JUMP'.
-		if op.arg != nil {
-			push := newInstruction(doc, op, "PUSH")
-			prog = append(prog, push)
-		}
-
-	default:
-		if _, ok := inst.opcode(); !ok {
-			return prog, fmt.Errorf("%w %s", ecUnknownOpcode, inst.op)
-		}
-		if op.arg != nil {
-			return prog, ecUnexpectedArgument
-		}
-	}
-
-	return append(prog, inst), nil
-}
-
-// validateJumpArg checks that argument to JUMP is a defined label.
-func (c *Compiler) validateJumpArg(doc *document, arg astExpr) error {
-	if arg == nil {
-		return nil // no argument is fine.
-	}
-	lref, ok := arg.(*labelRefExpr)
-	if !ok {
-		return ecJumpNeedsLiteralLabel
-	}
-	if lref.dotted {
-		return fmt.Errorf("%w %v", ecJumpToDottedLabel, lref)
-	}
-
-	var li *labelDefInstruction
-	if lref.global {
-		li = c.globals.label[lref.ident]
-	} else {
-		li, _ = doc.lookupLabel(lref)
-	}
-	if li == nil {
-		return fmt.Errorf("%w %v", ecJumpToUndefinedLabel, lref)
-	}
-	return nil
-}
-
-// expand appends the output of a macro call to the program.
-func (inst *macroCallInstruction) expand(c *Compiler, doc *document, prog []*instruction) ([]*instruction, error) {
-	name := inst.ident.text
-
-	var def *instructionMacroDef
-	var defdoc *document
-	if isGlobal(name) {
-		def, defdoc = c.globals.lookupInstrMacro(name)
-	} else {
-		def, defdoc = doc.lookupInstrMacro(name)
-	}
-	if def == nil {
-		return nil, fmt.Errorf("%w %%%s", ecUndefinedInstrMacro, name)
-	}
-
-	// Prevent recursion.
-	if !c.enterMacro(def) {
-		return nil, fmt.Errorf("%w %%%s", ecRecursiveCall, name)
-	}
-	defer c.exitMacro(def)
-
-	if len(inst.args) != len(def.params) {
-		return nil, fmt.Errorf("%w, macro %%%s needs %d", ecInvalidArgumentCount, name, len(def.params))
-	}
-	args := make(map[string]instrMacroArg)
-	for i, param := range def.params {
-		args[param] = instrMacroArg{calldoc: doc, expr: inst.args[i]}
-	}
-
-	// Expand. Here we clone the macro document and then run it.
-	macroDoc := *def.body
-	macroDoc.parent = defdoc
-	macroDoc.creation = inst
-	macroDoc.instrMacroArgs = args
-
-	prog = c.expand(&macroDoc, def.body.instructions, prog)
-	return prog, nil
-}
-
-func (c *Compiler) enterMacro(m *instructionMacroDef) bool {
-	if _, onStack := c.macroStack[m]; onStack {
-		return false
-	}
-	c.macroStack[m] = struct{}{}
-	return true
-}
-
-func (c *Compiler) exitMacro(m *instructionMacroDef) {
-	delete(c.macroStack, m)
-}
-
 // processIncludes reads all #included documents.
-func (c *Compiler) processIncludes(doc *document, stack []astStatement) {
+func (c *Compiler) processIncludes(doc *ast.Document, stack []ast.Statement) {
 	errs := c.globals.registerDefinitions(doc)
 	c.addErrors(errs)
 
-	var list []*includeInstruction
-	for _, inst := range doc.instructions {
-		inc, ok := inst.(*includeInstruction)
+	var list []*ast.IncludeSt
+	for _, inst := range doc.Statements {
+		inc, ok := inst.(*ast.IncludeSt)
 		if !ok {
 			continue
 		}
-		file, err := resolveRelative(doc.file, inc.filename)
+		file, err := resolveRelative(doc.File, inc.Filename)
 		if err != nil {
 			c.addError(inst, err)
 			continue
@@ -396,19 +203,18 @@ func (c *Compiler) processIncludes(doc *document, stack []astStatement) {
 		if incdoc == nil {
 			continue // there were parse errors
 		}
-		doc.includes[inc] = incdoc
+		c.includes[inc] = incdoc
 		list = append(list, inc)
 	}
 
 	// Process includes in macros.
-	for _, name := range sortedKeys(doc.instrMacros) {
-		m := doc.instrMacros[name]
-		c.processIncludes(m.body, append(stack, m))
+	for _, m := range doc.InstrMacros() {
+		c.processIncludes(m.Body, append(stack, m))
 	}
 
 	// Recurse.
 	for _, inst := range list {
-		incdoc := doc.includes[inst]
+		incdoc := c.includes[inst]
 		c.processIncludes(incdoc, append(stack, inst))
 	}
 }
@@ -421,7 +227,7 @@ func resolveRelative(basepath string, filename string) (string, error) {
 	return res, nil
 }
 
-func (c *Compiler) parseIncludeFile(file string, inst *includeInstruction, depth int) *document {
+func (c *Compiler) parseIncludeFile(file string, inst *ast.IncludeSt, depth int) *ast.Document {
 	if c.fsys == nil {
 		c.addError(inst, ecIncludeNoFS)
 		return nil
@@ -436,181 +242,32 @@ func (c *Compiler) parseIncludeFile(file string, inst *includeInstruction, depth
 		c.addError(inst, err)
 		return nil
 	}
-	p := newParser(file, content, c.lexDebug)
-	doc, errors := p.parse()
+	p := ast.NewParser(file, content, c.lexDebug)
+	doc, errors := p.Parse()
 	for _, e := range errors {
 		c.addError(inst, e)
 	}
 	if len(errors) > 0 {
 		return nil
 	}
-	doc.creation = inst
+	// Note that included documents do NOT have the including document set as Parent.
+	// The parent relationship is used during lookup of labels, macros, etc. and
+	// such definitions should not be shared between include files.
+	//
+	// Included documents do have a Creation though.
+	doc.Creation = inst
 	return doc
 }
 
-// expand of #include appends the included file's instructions to the program.
-// Note this accesses the documents parsed by processIncludes.
-func (inst *includeInstruction) expand(c *Compiler, doc *document, prog []*instruction) ([]*instruction, error) {
-	incdoc := doc.includes[inst]
-	if incdoc == nil {
-		return prog, nil // There was a parse error.
-	}
-	prog = c.expand(incdoc, incdoc.instructions, prog)
-	return prog, nil
-}
-
-// expand of #assemble performs compilation of the given assembly file.
-func (inst *assembleInstruction) expand(c *Compiler, doc *document, prog []*instruction) ([]*instruction, error) {
-	subc := NewCompiler(c.fsys)
-	subc.SetIncludeDepthLimit(c.maxIncDepth)
-	subc.SetMaxErrors(math.MaxInt)
-
-	file, err := resolveRelative(doc.file, inst.filename)
-	if err != nil {
-		return prog, err
-	}
-	bytecode := c.CompileFile(file)
-	if len(c.Errors()) > 0 {
-		return prog, nil
-	}
-	datainst := &instruction{data: bytecode}
-	return append(prog, datainst), nil
-}
-
-var zero = new(big.Int)
-
-// assignInitialPushSizes sets the pushSize of all PUSH and PUSH<n> instructions.
-// Arguments are pre-evaluated in this compilation step if they contain no label references.
-func (c *Compiler) assignInitialPushSizes(e *evaluator, prog []*instruction) {
-	for _, inst := range prog {
-		argument := inst.pushArg()
-		if argument == nil {
-			continue
-		}
-		inst.pushSize = 1
-		if s, ok := inst.explicitPushSize(); ok {
-			inst.pushSize = s
-		}
-
-		// Pre-evaluate argument.
-		env := newEvalEnvironment(inst.doc)
-		v, err := argument.eval(e, env)
-		var labelErr unassignedLabelError
-		if errors.As(err, &labelErr) {
-			// Expression depends on label position calculation, leave it for later.
-			continue
-		}
-		inst.argNoLabels = true
-		if err != nil {
-			c.addError(inst.ast, err)
-			continue
-		}
-		if err := c.assignPushArg(inst, v, true); err != nil {
-			c.addError(inst.ast, err)
-			continue
-		}
-	}
-}
-
-// computePC assigns the PC values of all instructions and labels.
-func (c *Compiler) computePC(e *evaluator, prog []*instruction) {
-	var pc int
-	for _, inst := range prog {
-		if li, ok := inst.ast.(*labelDefInstruction); ok {
-			e.setLabelPC(inst.doc, li, pc)
-		}
-
-		inst.pc = pc
-		size := 0
-		if inst.op != "" {
-			size = 1
-		}
-		if isPush(inst.op) {
-			size += inst.pushSize
-		} else {
-			size += len(inst.data)
-		}
-		pc += size
-	}
-}
-
-// assignArgs computes the argument values of all push instructions.
-func (c *Compiler) assignArgs(e *evaluator, prog []*instruction) (inst *instruction, err error) {
-	for _, inst := range prog {
-		if inst.argNoLabels {
-			continue // pre-calculated
-		}
-		argument := inst.pushArg()
-		if argument == nil {
-			continue // no arg
-		}
-		env := newEvalEnvironment(inst.doc)
-		v, err := argument.eval(e, env)
-		if err != nil {
-			return inst, err
-		}
-		if err := c.assignPushArg(inst, v, false); err != nil {
-			return inst, err
-		}
-	}
-	return nil, nil
-}
-
-// assignPushArg sets the argument value of an instruction to v. The byte size of the
-// value is checked against the declared "PUSH<n>" data size.
-//
-// If setSize is true, the pushSize of variable-size "PUSH" instructions will be assigned
-// based on the value.
-func (c *Compiler) assignPushArg(inst *instruction, v *big.Int, setSize bool) error {
-	if v.Sign() < 0 {
-		return ecNegativeResult
-	}
-	b := v.Bytes()
-	if len(b) > 32 {
-		return ecPushOverflow256
-	}
-	// TODO: also handle negative int
-
-	_, hasExplicitSize := inst.explicitPushSize()
-	if setSize && !hasExplicitSize {
-		inst.pushSize = c.autoPushSize(b)
-	}
-	if len(b) > inst.pushSize {
-		if !hasExplicitSize {
-			return ecVariablePushOverflow
-		}
-		return ecFixedSizePushOverflow
-	}
-
-	// Store data padded.
-	inst.data = make([]byte, inst.pushSize)
-	copy(inst.data[len(inst.data)-len(b):], b)
-	return nil
-}
-
-func (c *Compiler) autoPushSize(value []byte) int {
-	if len(value) > 32 {
-		panic("value too big")
-	}
-	if len(value) == 0 {
-		if c.usePush0 {
-			return 0
-		} else {
-			return 1
-		}
-	}
-	return len(value)
-}
-
 // generateOutput creates the bytecode. This is also where instruction names get resolved.
-func (c *Compiler) generateOutput(prog []*instruction) []byte {
+func (c *Compiler) generateOutput(prog *compilerProg) []byte {
 	if len(c.errors) > 0 {
 		return nil
 	}
 	var output []byte
-	for i, inst := range prog {
+	for _, inst := range prog.iterInstructions() {
 		if len(output) != inst.pc {
-			panic(fmt.Sprintf("BUG: instruction %d has pc=%d, but output has size %d", i, inst.pc, len(output)))
+			panic(fmt.Sprintf("BUG: instruction pc=%d, but output has size %d", inst.pc, len(output)))
 		}
 		if inst.op != "" {
 			opcode, ok := inst.opcode()
@@ -623,15 +280,4 @@ func (c *Compiler) generateOutput(prog []*instruction) []byte {
 		output = append(output, inst.data...)
 	}
 	return output
-}
-
-// opcode returns the EVM opcode of the instruction.
-func (inst *instruction) opcode() (evm.OpCode, bool) {
-	if isPush(inst.op) {
-		if inst.pushSize > 32 {
-			panic("BUG: pushSize > 32")
-		}
-		return evm.PUSH1 + evm.OpCode(inst.pushSize-1), true
-	}
-	return evm.OpByName(inst.op)
 }
