@@ -1,6 +1,6 @@
 // Code generated by "stringer -type tokenType"; DO NOT EDIT.
 
-package asm
+package ast
 
 import "strconv"
 
@@ -24,25 +24,16 @@ func _() {
 	_ = x[openParen-13]
 	_ = x[closeParen-14]
 	_ = x[comma-15]
-	_ = x[arithPlus-16]
-	_ = x[arithMinus-17]
-	_ = x[arithMul-18]
-	_ = x[arithDiv-19]
-	_ = x[arithMod-20]
-	_ = x[arithLshift-21]
-	_ = x[arithRshift-22]
-	_ = x[arithAnd-23]
-	_ = x[arithOr-24]
-	_ = x[arithHat-25]
-	_ = x[directive-26]
-	_ = x[instMacroIdent-27]
-	_ = x[openBrace-28]
-	_ = x[closeBrace-29]
+	_ = x[directive-16]
+	_ = x[instMacroIdent-17]
+	_ = x[openBrace-18]
+	_ = x[closeBrace-19]
+	_ = x[arith-20]
 }
 
-const _tokenType_name = "eoflineStartlineEndinvalidTokenidentifierdottedIdentifiervariableIdentifierlabelRefdottedLabelReflabeldottedLabelnumberLiteralstringLiteralopenParencloseParencommaarithPlusarithMinusarithMularithDivarithModarithLshiftarithRshiftarithAndarithOrarithHatdirectiveinstMacroIdentopenBracecloseBrace"
+const _tokenType_name = "eoflineStartlineEndinvalidTokenidentifierdottedIdentifiervariableIdentifierlabelRefdottedLabelReflabeldottedLabelnumberLiteralstringLiteralopenParencloseParencommadirectiveinstMacroIdentopenBracecloseBracearith"
 
-var _tokenType_index = [...]uint16{0, 3, 12, 19, 31, 41, 57, 75, 83, 97, 102, 113, 126, 139, 148, 158, 163, 172, 182, 190, 198, 206, 217, 228, 236, 243, 251, 260, 274, 283, 293}
+var _tokenType_index = [...]uint8{0, 3, 12, 19, 31, 41, 57, 75, 83, 97, 102, 113, 126, 139, 148, 158, 163, 172, 186, 195, 205, 210}
 
 func (i tokenType) String() string {
 	if i >= tokenType(len(_tokenType_index)-1) {
